cmd/link: allow deriving GNU build ID from action ID
While it is possible to embed a GNU build ID into the linked executable
by passing `-B 0xBUILDID` to the linker, the build ID will need to be
precomputed by the build system somehow. This makes it unnecessarily
complex to generate a deterministic build ID as it either requires the
biuld system to hash all inputs manually or to build the binary twice,
once to compute its hash and once with the GNU build ID derived from
that hash. Despite being complex, it is also inefficient as it requires
the build system to duplicate some of the work that the Go linker
already performs anyway.

Introduce a new argument "actionid" that can be passed to `-B` that
causes the linker to automatically derive the build ID from the Go build
ID's action ID, which is derived from all build inputs. This should thus
both be deterministic while the value also changes whenever the build
inputs change, which is the desired behaviour for the GNU build ID.

Furthermore, given that the `-B` flag currently requires a "0x" prefix
for all values passed to it, using "actionid" as value is a backwards
compatible change.

Fixes #41004